+++ src/app/[locale]/pomodoro/page.tsx
@@ -1,0 +1,7 @@
+import PomodoroContent from "@/components/contents/pomodoro";
+
+export default function PomodoroPage(){
+     return (
+          <PomodoroContent/>
+     )
+}