+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Calm Mood will be documented here.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## 2.0.2 - 2026-01-28
+### Added
+- Pomodoro Timer
+- New Translations for Pomodoro timer
+
 ## 2.0.1 - 2026-01-11
 ### Added
 - Form to set the interval time